refactor: 누락 어노테이션 추가
