+++ src/main/java/com/example/tripKo/domain/place/restaurant/PlaceRestaurant.java
@@ -1,15 +1,21 @@
-package com.example.tripKo.place.restaurant;
+package com.example.tripKo.domain.place.restaurant;
 
 import com.example.tripKo.BaseTimeEntity;
-import com.example.tripKo.place.Category;
-import com.example.tripKo.place.Place;
+import com.example.tripKo.domain.place.Category;
+import com.example.tripKo.domain.place.Place;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.DynamicInsert;
+import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
 
+import static javax.persistence.FetchType.LAZY;
+
+@DynamicInsert
+@DynamicUpdate
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
@@ -40,7 +46,8 @@ public class PlaceRestaurant extends BaseTimeEntity {
     @Column(length = 100)
     private String holidayDate;
 
-    @OneToOne
+    @OneToOne(fetch = LAZY)
+    @JoinColumn(name = "place_id", nullable = false)
     private Place place;
 
     @Builder
